improve handling of deleted branches.
In the scenario where a branch is created, but is then deleted and
re-branched, we need to be smarter.

The way we handle this now is that we treat each generation of a branch
as a distinct branch entity during analysis.

During completion, any dead branches are created in a fixup area so that
they will not appear under that branch name in the final repo.
The fixup area is deleted at the end of the run.

Deleted tags are no longer represented in the git repo.